world_transformation aufgabe 3

+++ framework/Sphere.hpp
@@ -2,7 +2,7 @@
 #define BUW_SPHERE_HPP
 
 #include "Shape.hpp"
-#include <glm/vec3.hpp>
+#include <glm/glm.hpp>
 #include "color.hpp"
 #include "Ray.hpp"
 #include "intersection.hpp"
@@ -10,12 +10,12 @@
 class Sphere : public Shape{
 	public:
 		Sphere();
-		Sphere(glm::vec3 const& center, float radiusr, std::string name, std::shared_ptr<Material> const& material);
+		Sphere(glm::dvec3 const& center, float radiusr, std::string name, std::shared_ptr<Material> const& material);
 
 		/* virtual  float area() const ;
 		virtual  float volume() const override;*/
 
-		glm::vec3 getcenter() const;
+		glm::dvec3 getcenter() const;
 		void setcenter_x(double centerx) &;
 		void setcenter_y(double centery) &;
 		void setcenter_z(double centerz) &;
@@ -26,12 +26,12 @@ class Sphere : public Shape{
 
 		/* virtual */ void print(std::ostream&) const override;
 
-		glm::vec3 kreisNormale(glm::vec3 const& mittelpunkt, glm::vec3 const& surface) const;
+		glm::dvec3 kreisNormale(glm::dvec3 const& mittelpunkt, glm::dvec3 const& surface) const;
 		
 		~Sphere();
 
 	private:
-		glm::vec3 center_;
+		glm::dvec3 center_;
 		float radius_;
