+++ framework/Triangle.hpp
@@ -2,7 +2,7 @@
 #define BUW_TRIANGLE_HPP
 
 #include <array>
-#include <glm/vec3.hpp>
+#include <glm/glm.hpp>
 #include <glm/gtx/intersect.hpp>
 #include "Shape.hpp"
 #include "Ray.hpp"
@@ -14,16 +14,16 @@ class Triangle : public Shape
 	
 		public:
 			Triangle();
-			Triangle(glm::vec3 const& u, glm::vec3 const& v, glm::vec3 const& w);
-			glm::vec3 v1 () const;
-			glm::vec3 v2 () const;
-			glm::vec3 v3 () const;
+			Triangle(glm::dvec3 const& u, glm::dvec3 const& v, glm::dvec3 const& w);
+			glm::dvec3 v1 () const;
+			glm::dvec3 v2 () const;
+			glm::dvec3 v3 () const;
 
 			Intersection intersect (Ray const& ray) const override;
-			glm::vec3 barycentric_to_world (glm::vec3 const& bary) const;
+			glm::dvec3 barycentric_to_world (glm::dvec3 const& bary) const;
 
 		private:
-			glm::vec3 v1_, v2_, v3_, n_;
+			glm::dvec3 v1_, v2_, v3_, n_;
 			std::string name_;
 			std::shared_ptr<Material> material_;
 };
